+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!--next-version-placeholder-->
 ## [v3.8.15-dev3] - 2026-04-21
 ### Added
-* DICOMPlugin: Document data types and processes
+* DICOMPlugin: Document data types and processes.
 
 ## [v3.8.15-dev2] - 2026-04-14
 ### Fixed